Create Password generaor in python
The Password Generator script in Python generates a random, secure password using a combination of uppercase and lowercase letters, digits, and special characters. It uses the random.choice() function to select characters randomly from the predefined set of characters. By default, the generated password is 12 characters long, but this can be customized by passing a desired length to the generate_password() function. This script helps in creating strong, unpredictable passwords for secure applications.